requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-jose
argon2-cffi
python-dotenv
pdfplumber
python-docx
